backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
textblob==0.17.1
python-dateutil==2.8.2
bcrypt==4.0.1
google-generativeai==0.3.1
langchain>=0.1.0
langchain-community>=0.0.10
